Readme.md
Toward the end of 2021, we still do not see any sign of Covid-19 cases slowing down in Malaysia. Asides from vaccination, is there other measures that can mitigate the COVID-19? For me is AI. AI is the new electricity that can transform business and healthcare into new form just like electricity did 100 years ago. The core of AI is machine learning. Can the AI transform lockdown process to have the minimum impact on the economic? To show how it works, I have written hundred lines of machine learning codes to work with Mysejahtera together.
Two problem framing are formed. The first problem framing is “Given the attributes of each user, predict which individual is the next cluster contributor”. The second problem framing is “Given the locations each user visited, predict which location is the next cluster contributor”. 
For the first task, a simple data frame is created with attributes such as gender, age, health condition, close contact and etc. Which in practical can be collected by the Mysejahtera app. With this information, the data are pre-processed, such as to label encode all the categorical attributes to fit into machine learning models.
As a data scientist, I have built my own data modelling tools to train the dataset to learn the best hypothesis mapping. For example, the code for decision tree models can produce 1680 different combinations of decision tree to find the best parameters and best accuracy. After training the data, naïve bayes model is not bad. So the naïve bayes is used to test the unseen data. And the results show that it has 100% accuracy.  
From the result, the flagged individual is Farah, who is the next contributor suggested by the machine learning. So, the result is processed and the instances are pulled into a new form to deploy and visualise. After that, the next question is - How to automate the movement control process? To answer this, the machine learning result can be worked together with Mysejahtera, to automate the process. For example, the dashboard will show a warning letter for not allow to enter. If the individual scan during the quarantine period, a link will be created to submit for a punishment. By this, it can automate the punishment. But how is a fair punishment for violations? For new learner, it is allowed to have 3 times before submit a punishment and different punishment value for different economic class.
Lastly, the results are deployed in a web app to visualise the status of the individual. And the first task demonstration finished.
The second task is using unsupervised learning. To recall the problem framing, given the locations each user visited, predict which location is the next cluster contributor. Apriori algorithm will be used to deployed the association rules. To show how it works, a simple data frame is created, with attributes such as USER ID, Date, LOCATIONS visited are formed. The attributes can be collected by the mysejahtera app. Then the data is pre-processed into a long format with label encoded 0 and 1. Then association rules mining from Apriori algorithm is apply and fit to the data. The result show that ['SHOP_XX4', 'SHOP_XX3'] are the WARM BED of COVID-19.
Then the next question is, how to automate the lockdown process for the flagged business?  Again, Mysejahtera app can be worked together with the machine learning models, an automate notification will be scan to the contact person of the business. Whenever individual has scan the qr code of the flagged business, not only individual will be notified that the business is being in a lock down process and return a not allowed to enter message.
Then, the results of the unsupervised learning is deployed into a web app. So that it is to let known to law enforcement to make sure that the businesses are being lock down. So, if let say in practical there are 1000 business are red flagged, are the law enforcement need to go and check every day? The answer is no. It can again use machine learning to predict which business based on the attributes such as business type, opening hours, region and etc. using unsupervised learning clustering technique. So that the law enforcement can accurately identify which business has violated the SOP. Thanks for listening.
So by using AI work together with MySejahtera App, it will auto implement a smart and accurate lockdown to mitigate the COVID-19 spreading with the least impact toward the Economic